Add COM helpers
